Removed Procedural Map Gen
Easily enough added back in, but I don't need it.

Added functionality for enemy spawning to not spawn off grid, and to spawn off screen at a given range.

Enemies are now in an Enemies object, as to not clutter the hierarchy. Projectiles are now under their shooter's object too.