Character Selection Working
The character selection now works, and has primitive UI that allows the player to see the changes made.

Text shows up on the right that indicates whether changing from the selected class to another is good or bad for stats X Y and Z using comparisons, and is shown with coloured text.