Replaced default font with Norse font
Not fully done but it looks less -default?- and had to change the spacing and sizing.

Also fixed incorrectly assigned isActive on panels in the level up scene.

I don't think the upgrade stats panel showed up before after sorting out the abilities and it does now, so that's good.