Fixed Enemy Anim and Knockback Bugs
Knockback was applying a force to the enemies such that they were rotating on the Z axis.

Only attack and die enemy anims were working. Walking and idling were not. That is now fixed.

Enemy projectile sprite also changed to the arrow from the goblin sprite set. Projectile code changed so that the arrow points in the right direction.