Fixed Wave Progression Bug
Game was not letting me complete a wave. Enemy GO did not implement IPooledObject, meaning each Enemy's spawner field was null, so the enemy couldn't tell the spawner, to tell the wavemanager, to add it to the death toll.

Also changed alpha value of rarity panel in level up scene to match that of the abilities panel.