Adding Snowy Trees
Snowy trees have a prefab for each type. There's sprites for 5 colours for each tree but that's a lot of work and really it can be done through changing the Color in the spriteRenderer.

There is a core prefab and different variants. There is a master controller and overrides for the anims, each with a clip they override with.

Made it so props are randomised to be between a little smaller than before and a little less big.

Made tilemap bigger, and the outskirts have thicker grass as a "don't go further" sort of thing". Need to replace with snow really.

Snowy trees are all hooked up with animations, sprite sorters and colliders.

Snowy trees are also now the only type of tree or plant in the main game, but assets remain for non-snowy ones.

Added some scripts to batch slice etc. Also added Python scripts for the future but they aren't particularly generic.

Don't really know the best way to add new sprites without spending hours.