Weather Controller
Weather can now be controlled by one script. Hopefully it'll make it easier to mess with the weather when using abilities.

Although, a non particle system approach could be better, so I can do, say a Hammer Slam, and it push the wind/fog away from my for a second.