Added Shockwave and Dust Cloud Art
Updated ArtCredits too.

Hooked the prefabs up for radial dust cloud and shockwave to the hammer slam ability.

Added script for FadeAndDestroy that lives on the prefabs (can be used for other things), and can toggle whether or not there is a fade, when it starts, how long it fades for, etc.

Added hammerSlam position inside the player object and class so the hammrerSlamAbility can use that rather than hard coding it in. Added reference inside Player as well for the weaponHolder and reflected these changes in AttackManager.